+++ packages/react/components/box/context/boxContext.tsx
@@ -1,0 +1,3 @@
+import React from 'react'
+
+export const BoxContext = React.createContext({ fullHeight: false })