+++ packages/react/components/input/Input.tsx
@@ -202,7 +202,7 @@ const Input = (
           onBlur={handleBlur}
         />
         {hasIcon && !localStatus && !loading && <IconWrapper name={iconNameLeft as unknown as IconName} />}
-        {iconNameLeft && !loading && <IconWrapper className={'icon-left'} name={iconNameLeft as unknown as IconName} />}
+        {iconNameLeft && <IconWrapper className={'icon-left'} name={iconNameLeft as unknown as IconName} />}
         {iconNameRight && !loading && type !== 'password' && (
           <IconWrapper className={'icon-right'} name={iconNameRight as unknown as IconName} />
         )}