+++ packages/react/components/progress/index.ts
@@ -1,5 +1,5 @@
-import Progress from './Progress'
-import ProgressRadial from './radial'
-import ProgressRadialItem from './radial/item/ProgressRadialtem'
+import Progress from '@/components/progress/Progress'
+import ProgressRadial from '@/components/progress/radial'
+import ProgressRadialItem from '@/components/progress/radial/item/ProgressRadialtem'
 
 export { Progress, ProgressRadial, ProgressRadialItem }