+++ packages/react/components/sticker/index.ts
@@ -1,5 +1,3 @@
 import Sticker from './Sticker'
 
-export * from './StickerEnum'
-
 export { Sticker }