+++ CHANGELOG.md
@@ -3,6 +3,35 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.0.5] - 2025-01-30
+- Fix Native Columns
+- Fix Scrollable Columns
+- Fix CheckboxTiles Native
+- Update Link interface
+- Fix List style
+- Fix Section background src
+- Fix font color inside Container
+- Fix AccordionBody Width Native
+
+## [4.0.4] - 2025-01-14
+### Fix
+- Add npmrc
+
+## [4.0.3] - 2025-01-13
+### Fix
+- Fix [List] Native padding size
+- Fix [Columns] Native use column as composant
+- Update [View] Add markup's
+- Update [Tag] Native Icons spacing
+
+## [4.0.2] - 2025-01-08
+### Fix
+- Fix [Price] Native when "0" amount
+- Fix [Divider] Native content color
+- Fix [Timeline] Native default color
+- Fix [List] Native background
+- Fix [Stepper] Native fullwidth & crash
+- Fix [Columns] Native no size on column / narrow
 
 ## [4.0.1] - 2024-12-16
 ### Remove